Fix the Travis status image and put it to the right of the repo's title
So, apparently somebody made a derp and pointed our Travis image to the wrong
repository.... Fixed. Also, GitHub people tend to put the Travis status to the
right of the project title in the readme, and link it in such a way that
clicking on the image will bring you to the repo's Travis page, not the status
image itself. Also fixed.